+++ Assets/MirrorMarketSDK/Runtime/MirrorWrapper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 306f69768d6114f989d144fd22df1b13
+guid: d65e8e09021e54afbbf5a3104b5b433d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
