--- Assets/MirrorMarketSDK/Runtime/Services/Implementations/MirrorSDKAuth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d6d96129a4e642de99195056ad3ef2d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
