--- Assets/MirrorMarketSDK/Runtime/Services/Implementations/MirrorSDKMarket.cs
@@ -1,5 +1,0 @@
-﻿using System;
-public partial class MirrorSDK
-{
-    //public 
-}
