+++ Assets/MirrorMarketSDK/Runtime/Services/Implementations/MirrorWrapperAuth.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 03113e111863a4fd68539598dc2dbb97
+guid: a23e57bf58b68433886db196c0be5608
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
