+++ Assets/MirrorMarketSDK/Runtime/Services/Implementations/MirrorWrapperMarket.cs
@@ -1,0 +1,9 @@
+﻿using System;
+namespace MirrorworldSDK.Wrapper
+{
+    public partial class MirrorWrapper
+    {
+        //public 
+    }
+}
+
