+++ Assets/MirrorMarketSDK/Runtime/Services/Implementations/MirrorWrapperMarket.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d887a818e74eb4316a839030f930036b
+guid: 75d122af411734317a109e756abc3384
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
