--- Assets/MirrorMarketSDK/Runtime/Services/Implementations/NewClass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a97a5bfb68d24d2dbbf8eaf48edea03
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
