+++ Assets/MirrorMarketSDK/Runtime/Utils/Enums.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 75476b67117c64ab1a25a2b94db96dcb
+guid: 3171562e4652d4c2eb089afd64df7a12
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
