+++ Assets/Test/Test.cs
@@ -1,6 +1,5 @@
 
 using MirrorworldSDK;
-using ThinkingAnalytics;
 using UnityEngine;
 
 public class Test : MonoBehaviour
@@ -8,7 +7,7 @@ public class Test : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        ThinkingAnalyticsAPI.CalibrateTime(1212);
+
     }
 
     // Update is called once per frame
