--- Assets/ThinkingAnalytics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 60b198fb61dd042d5b21af809183759f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
