--- Assets/ThinkingAnalytics/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1c3596aacf52e4dce91260e35b40c684
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
