--- Assets/ThinkingAnalytics/Editor/TD_Inspectors.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9e175a5169e2c4de78d15b28f2ce6520
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
