--- Assets/ThinkingAnalytics/Exception.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f347e0b07c5d64926ac64031c3bba013
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
