--- Assets/ThinkingAnalytics/Exception/ThinkingSDKException.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76f285e49c9584fce99e9fe1536fea17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
