--- Assets/ThinkingAnalytics/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4e5d07cff2f9b45d6afa9ab9a879fbff
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
