--- Assets/ThinkingAnalytics/Resources/ta_public_config.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- ThinkingAnalytics DisablePresetProperties start -->
-    <!-- <string-array name="TDDisPresetProperties"> -->
-       <!-- <item>#disk</item> -->
-       <!-- <item>#fps</item> -->
-       <!-- <item>#ram</item> -->
-       <!-- <item>#app_version</item> -->
-       <!-- <item>#os_version</item> -->
-       <!-- <item>#manufacturer</item> -->
-       <!-- <item>#device_model</item> -->
-       <!-- <item>#screen_height</item> -->
-       <!-- <item>#screen_width</item> -->
-       <!-- <item>#carrier</item> -->
-       <!-- <item>#device_id</item> -->
-       <!-- <item>#system_language</item> -->
-       <!-- <item>#lib</item> -->
-       <!-- <item>#lib_version</item> -->
-       <!-- <item>#os</item> -->
-       <!-- <item>#bundle_id</item> -->
-       <!-- <item>#install_time</item> -->
-       <!-- <item>#start_reason</item> -->
-       <!-- <item>#simulator</item> -->
-       <!-- <item>#network_type</item> -->
-       <!-- <item>#start_reason</item> -->
-       <!-- <item>#resume_from_background</item> -->
-       <!-- <item>#title</item> -->
-       <!-- <item>#screen_name</item> -->
-       <!-- <item>#url</item> -->
-       <!-- <item>#referrer</item> -->
-       <!-- <item>#element_type</item> -->
-       <!-- <item>#element_id</item> -->
-       <!-- <item>#element_position</item> -->
-       <!-- <item>#element_content</item> -->
-       <!-- <item>#element_selector</item> -->
-       <!-- <item>#app_crashed_reason</item> -->
-    <!-- </string-array> -->
-    <!-- ThinkingAnalytics DisablePresetProperties end -->
-
-    <!-- ThinkingAnalytics disable C# Exception -->
-    <!-- <bool name="DisableCSharpException">true</bool> -->
-</resources>
