--- Assets/ThinkingAnalytics/Resources/ta_public_config.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b67db39cd532b4f08b95b2cf5938553f
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
