--- Assets/ThinkingAnalytics/ThinkingAnalytics.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 986ad473b78d14e08b1f05d368cb1591
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
