--- Assets/ThinkingAnalytics/ThinkingAnalyticsAPI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d7df5c5e5307f460e917e72bf75ec567
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
