--- Assets/ThinkingAnalytics/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: efe5e95f8073844fcbaedf8453895210
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
