--- Assets/ThinkingAnalytics/Utils/TD_Log.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b49f9a69526c4417963783322d7e904
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
