--- Assets/ThinkingAnalytics/Utils/TD_MiniJson.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 03cd5f56606ff4dd694406f5594f0cf2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
