--- Assets/ThinkingAnalytics/Utils/TD_PublicConfig.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 31781ef5586fb47b68b6d0dbb4bc79a8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
