--- Assets/ThinkingAnalytics/Wrapper.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 57bb4a9644e4f47f48024ccf8af3de0c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
