--- Assets/ThinkingAnalytics/Wrapper/TD_PCWrapper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f9ded317afe443d894852ab82fad3ad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
