+++ UserSettings/EditorUserSettings.asset
@@ -5,6 +5,9 @@ EditorUserSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 4
   m_ConfigSettings:
+    RecentlyUsedSceneGuid-0:
+      value: 5a070400570d0b085c5b59744377064317151c72297f74657e2d4a63b7e2306d
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
